🧪 Support registering custom api plugin

+++ spkl/SparkleXrm.Tasks/PluginRegistraton.cs
@@ -15,7 +15,11 @@ namespace SparkleXrm.Tasks
         private OrganizationServiceContext _ctx;
         private IOrganizationService _service;
         private ITrace _trace;
-
+        private static sdkmessageprocessingstep_stage[] SupportedPluginStages = new sdkmessageprocessingstep_stage[] {
+                sdkmessageprocessingstep_stage.Prevalidation, 
+                sdkmessageprocessingstep_stage.Preoperation,
+                sdkmessageprocessingstep_stage.Postoperation,
+                sdkmessageprocessingstep_stage.Postoperation_Deprecated };
         public PluginRegistraton(IOrganizationService service, OrganizationServiceContext context, ITrace trace)
         {
             _ctx = context;
@@ -307,17 +311,66 @@ namespace SparkleXrm.Tasks
 
                     foreach (var pluginAttribute in pluginAttributes)
                     {
-                        RegisterStep(sdkPluginType, existingSteps, pluginAttribute);
+                        var attribute = pluginAttribute.CreateFromData();
+                        if (attribute.Name == null && attribute.Message != null)
+                        {
+                            // Custom API registration
+                            RegisterCustomApi(sdkPluginType, existingSteps, attribute);
+                        }
+                        else
+                        {
+                            // Plugin Step registration
+                            RegisterStep(sdkPluginType, existingSteps, attribute);
+                        }
                     }
 
                     // Remove remaining Existing steps
                     foreach (var step in existingSteps)
                     {
-                        _trace.WriteLine("Deleting step '{0}'", step.Name, step.Stage);
-                        _service.Delete(SdkMessageProcessingStep.EntityLogicalName, step.Id);
+
+                        if (SupportedPluginStages.Contains(step.Stage??sdkmessageprocessingstep_stage.InitialPreoperation_Forinternaluseonly))
+                        {
+                            _trace.WriteLine("Deleting step '{0}'", step.Name, step.Stage);
+                            _service.Delete(SdkMessageProcessingStep.EntityLogicalName, step.Id);
+                        }
                     }
                 }
             }
+        }
+
+        private void RegisterCustomApi(PluginType sdkPluginType, List<SdkMessageProcessingStep> existingSteps, CrmPluginRegistrationAttribute attribute)
+        {
+            // Register against a Custom Api 
+            // Check it exists using the unqiue message name
+            var existingApi = (from s in _ctx.CreateQuery<CustomAPI>()
+                         where s.UniqueName == attribute.Message
+                         select new CustomAPI()
+                         {
+                             Id = s.Id,
+                             PluginTypeId = s.PluginTypeId
+                         }).ToList();
+            
+            if (existingApi.Count>0)
+            {
+                // Update Api Registration if it's a different type Id or null
+                var customApi = existingApi.First();
+                if (customApi.PluginTypeId?.Id != sdkPluginType.Id)
+                {
+                    customApi.PluginTypeId = sdkPluginType.ToEntityReference();
+                    _service.Update(customApi);
+
+                   _trace.WriteLine($"Registered Plugin Type {sdkPluginType.Id} against Custom Api '{attribute.Message}'");
+                }
+                else
+                {
+                    _trace.WriteLine($"Plugin Type {sdkPluginType.Id} is already registered against Custom Api '{attribute.Message}'");
+                }
+
+            }
+            else
+            {
+                _trace.WriteLine($"Warning: Cannot find custom api with UniqueName '{attribute.Message}'");
+            }  
         }
 
         private List<SdkMessageProcessingStep> GetExistingSteps(PluginType sdkPluginType)
@@ -347,10 +400,8 @@ namespace SparkleXrm.Tasks
             return steps;
         }
 
-        private void RegisterStep(PluginType sdkPluginType, List<SdkMessageProcessingStep> existingSteps, CustomAttributeData pluginAttribute)
+        private void RegisterStep(PluginType sdkPluginType, List<SdkMessageProcessingStep> existingSteps, CrmPluginRegistrationAttribute pluginStep)
         {
-            var pluginStep = pluginAttribute.CreateFromData();
-
             SdkMessageProcessingStep step = null;
             Guid stepId = Guid.Empty;
             if (pluginStep.Id != null)
@@ -398,7 +449,10 @@ namespace SparkleXrm.Tasks
             step.Configuration = pluginStep.UnSecureConfiguration;
             step.Description = pluginStep.Description;
             step.Mode = pluginStep.ExecutionMode == ExecutionModeEnum.Asynchronous ? sdkmessageprocessingstep_mode.Asynchronous : sdkmessageprocessingstep_mode.Synchronous;
-            step.AsyncAutoDelete = pluginStep.ExecutionMode == ExecutionModeEnum.Asynchronous ? pluginStep.DeleteAsyncOperation : false;
+            if (pluginStep.ExecutionMode == ExecutionModeEnum.Asynchronous)
+            {
+                step.AsyncAutoDelete = pluginStep.DeleteAsyncOperation;
+            }
             step.Rank = pluginStep.ExecutionOrder;
             int stage = 10;
             switch (pluginStep.Stage)
@@ -435,7 +489,7 @@ namespace SparkleXrm.Tasks
             step.SdkMessageFilterId = sdkMessagefilterId != null ? new EntityReference(SdkMessageFilter.EntityLogicalName, sdkMessagefilterId.Value) : null;
             step.SdkMessageId = new EntityReference(SdkMessage.EntityLogicalName, sdkMessageId.Value);
             step.FilteringAttributes = pluginStep.FilteringAttributes;
-            step.AsyncAutoDelete = pluginStep.DeleteAsyncOperation;
+           
             if (step.Id == Guid.Empty)
             {
                 _trace.WriteLine("Registering Step '{0}'", step.Name);
